fixed first dpgs registry link
The first set of  G's put into the registry were hard coded. This was before a registry website. They do not conform to the current standard data structure. As such, they all point to one single blogpost page listing all of them.
This link to that  post from the registry list was broken as a result of a previous PR.